+++ spec/features/provider_spec.rb
@@ -46,4 +46,30 @@ RSpec.feature "Provider browsing" do
 
     expect(body).to_not have_content "Recently added resources"
   end
+
+  scenario "I can see 'Manage the resource' button if i am an admin provider" do
+    admin = create(:user)
+    dataAdmin = create(:data_administrator, first_name: admin.first_name, last_name: admin.last_name, email: admin.email)
+    provider = create(:provider, data_administrators: [dataAdmin])
+
+    checkin_sign_in_as(admin)
+
+    visit provider_path(provider)
+
+    expect(page).to have_link("Browse resources")
+    expect(page).to have_content("Manage the provider")
+  end
+
+  scenario "I cannnot see 'Manage the resource' button if i am not an admin provider" do
+    user, admin = create_list(:user, 2)
+    dataAdmin = create(:data_administrator, first_name: admin.first_name, last_name: admin.last_name, email: admin.email)
+    provider = create(:provider, data_administrators: [dataAdmin])
+
+    checkin_sign_in_as(user)
+
+    visit provider_path(provider)
+
+    expect(page).to have_link("Browse resources")
+    expect(page).to_not have_content("Manage the provider")
+  end
 end