+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Please view this file on the master branch, on stable branches it's out of date.
 
+## [3.17.0-milestone]
+
+### Added
+- Resource provider can navigate to the Provider Component from the Provider Presentation Page (@kmarszalek)
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [3.16.0-milestone]
 
 ### Added